"Change from .txt to .md files"
The commit is about renaming all the file extensions from .txt to .md and also removing the space component and replacing it with two line breaks ('<br>') in the markdown syntax. The need was to maintain a consistent and global file format (.md) which is superior for documentation as compared to .txt files. Also, since the newline character in markdown `<br>` is more standard and widely accepted, it was decided to replace the 'Space' component with it. The diff also shows the removal of a file named "Arty-test/main.txt" possibly because the content was not relevant or necessary anymore.